Added optimization (-O) and math library (-lm) flags.
The optimization flag is required to prevent a bug from an unused line of code in the netwk function. The result of the line is not used so the compiler removes it when optimizing. When compiling with debugging turned on (-g) the line causes a divide by zero exception. The same bug and fix occurs in the FORTRAN code as well (tested with g77).

The -lm flag is required to link the c math library wen compiling on linux.